requirements.txt
# pip install -r requirements.txt

streamlit~=1.2.0

# base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python-headless>=4.1.2
Pillow~=8.4.0
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
paddlepaddle >= 2.0.1

# logging -------------------------------------
tensorboard>=2.4.1
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas~=1.1.5

# export --------------------------------------
# coremltools>=4.1
# onnx>=1.8.1
# scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
thop~=0.0.31-2005241907
pycocotools>=2.0  # COCO mAP

opencv-python~=4.5.4.60
levenshtein~=0.16.0
requests~=2.26.0